Editor: SCI_LINESCROLL: treat columns as signed

+++ src/Editor.cxx
@@ -7414,7 +7414,7 @@ sptr_t Editor::WndProc(unsigned int iMessage, uptr_t wParam, sptr_t lParam) {
 
 	case SCI_LINESCROLL:
 		ScrollTo(topLine + lParam);
-		HorizontalScrollTo(xOffset + wParam * vs.spaceWidth);
+		HorizontalScrollTo(xOffset + static_cast<int>(wParam) * vs.spaceWidth);
 		return 1;
 
 	case SCI_SETXOFFSET:
